+++ src/base/Field.php
@@ -400,6 +400,14 @@ abstract class Field extends SavableComponent implements FieldInterface
     /**
      * @inheritdoc
      */
+    public function getInputId(): string
+    {
+        return Html::id($this->handle);
+    }
+
+    /**
+     * @inheritdoc
+     */
     public function useFieldset(): bool
     {
         return false;